refactor(database): Split `database.ts` into `config.ts` & `initDb.ts`
