--- database/src/database.ts
@@ -1,26 +1,0 @@
-import { Lesson, StudentFromList } from "@turbo-schedule/common";
-
-import low, { AdapterAsync } from "lowdb";
-import FileAsync from "lowdb/adapters/FileAsync";
-
-export const databaseFile: string = process.env.NODE_ENV === "test" ? "database.test.json" : "database.json";
-
-export interface DbSchema {
-	students: StudentFromList[];
-	lessons: Lesson[];
-}
-
-export const defaultDbState: DbSchema = { students: [], lessons: [] };
-
-export type Db = low.LowdbAsync<DbSchema>;
-
-const initDb = async (storageFile: string = databaseFile): Promise<Db> => {
-	const adapter: AdapterAsync<DbSchema> = new FileAsync<DbSchema>(storageFile);
-	const db: Db = await low(adapter);
-
-	db.defaults(defaultDbState);
-
-	return db;
-};
-
-export { initDb };
